move Kit DTO
